+++ docs/Changelog.md
@@ -1,9 +1,11 @@
 # Changelog
 
 # 0.2.20
-  - add `within` logic to timeframed pass-fail criterias 
-  - added `steps` option to execution settings.
-
+  - add `within` logic to timeframed pass-fail criterias
+  - added `use-dns-cache-mgr` option.
+  - default-domain option renamed to default-address (scheme, hostname, port are now parsed from this option).
+  - added `steps` option to execution settings
+  
 # 0.2.19
   - fixed bug when in distributed tests VU count was not calculated properly.
   - auto-append `${__machineName()}` for thread names in distributed test